Optimization for aliases
Optimized AliasAdd() and AliasGet() by avoiding the usage of
Tcl_SetVar2Ex() and Tcl_GetVar2Ex(). This change improves the
performance of AliasAdd() by ~20% and AliasGet() by 10%.
Strangely, Tcl does not provide an interface to the
Tcl_Obj-based TclObjUnsetVar2(), which is used internally
by Tcl_UnsetVar2()), such we could also avoid the latter.