Don't allow bytearrays as name of non-pos args
We try to address the problem, that the argument parser might add string reps to Tcl_Objs
without string reps. This is in particular nasty for pure byte arrays in Tcl 8.6, which
are defined as having no string. We do not want to change purtiy just by passing
such values in arguments,