- improved handling of object property autonamed
A call of "new" calls internally the "create" method. When the
"create" method is overloaded, we want to be able to check already
on this level, whether the object is autonamed or not. The previous
version has set the property at the end of the "new" method, which
was too late for ttrace.