Do not support predefined parameter options spelled with trailing content
This change makes argument type checking more strict by rejecting
names as synonyms which are longer than the specified values.
Background: Previous versions did not handle cases correctly where an
application type name starts with the same characters as a built-in
type. So it was not possible to define a type "object_id", since a a
spec of the form "x:object_id" was interpreted as "x:object".