Pack Tcl-version into name for Tcl 9.*
With Tcl 9, it is possible to install nsf for 8.7 and 9.0 parallel in
the same directories. The trick is naming the shared library
"libtcl9nsf2.4.0.so" instead of "libnsf2.4.0.so" when compiled with
Tcl 9.0 (keeping the original name when compiled with Tcl 8.x).
many thanks to Jan Nijtmans for the patch

Whitespace cleanup