Deactivate support of TIP 629 for the time being
When NS_TCL_HAVE_TIP629 (TIP supporting > 2^31 elements in object
vectors) is activated, Tcl defines different objProcs (Tcl_ObjCmdProc
and Tcl_ObjCmdProc2) for the old and new API, where the usage of these
depends on a cmdWrapperProc. Unfortunately, the resolving of these API
calls is performed through the wrapper via CmdWrapperInfo, which is
not exported. We have to think how to resolve these to make this
working as with prior Tcl versions.