Fixed warnings detected with branch "disabletcl8api"
The Tcl9 native interface is based on "Tcl_Size" instead of "int", but
the default mode hides these differences via some macro-magic. Typical
such cases are API calls like:

   Tcl_ListObjLength(interp, methodObj, &length);
   Tcl_ListObjGetElements(interp, methodObj, &oc, &ov)

The macro-magic queries the size of the provided parameters and uses
the old or new interface depending on it (while hiding some more
potential problems, a compiler can detect). The branch
"disabletcl8api" removes the macro-magic and reveals the potential
problem areas, where the code is still depending on the old 32bit
interface.

This commit fixes such cases and the collateral damage.