Provided a Tcl_DupInternalRepProc() function for Tcl_ObjType "nsfParam"
Recent versions of Tcl9 call Tcl_DupInternalRepProc() for Tcl_ObjType
"nsfParam", which was a placeholder function before, leading to a
Tcl_Abort(). The provided implementation provides now a basic logic
for "duplicating" internal representations of parameter lists based on
refcounts (somewhat similar to ProcBodyDup), but it has to be tested
further, with a proper test case.

The call of the Tcl_DupInternalRepProc function happens so far during
the startup of NaviServer compiled with Tcl9. It might be necessary to
overthink the competing logics of the "canDelete" member and the
reference count. For now, the intention is to get the NaviServer +
Tcl9 regression test stop from aborting.