Fixed handling of invalid parameter specs
This bug was occuring e.g. with an invalid parameter spec as in

   :property {fiscalyear:integer, required}

where an additional space was introduced. The bug was reported by
Maksym Zinchenko on the xotcl mailing list (many thanks for that!)
and lead to an attempt of adding an empty parameter option.
Now, an error is generated in this case.