fix for issue #4; Memory leak of NsfMethodContext with Tcl9
Ths is not a fully satisfying fix, since it just fixes the symptom, not the cause.
Also the old variant is supposed to run free of this issue. However, we are
talking here about a single Tcl_Obj inside the nx object system, so the harm
is very limited. More details in the ticket.